server/pyproject.toml
[project]
name = "pet-typeless-server"
version = "0.2.0"
description = "PetTypeless Relay Server — proxies client audio to 豆包 bigmodel_async ASR"
readme = { text = "", content-type = "text/markdown" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
pet-typeless-server = "pet_typeless_server.main:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
